board/MynaPlayer odyssey: removed initramfs.
This build no longer requires an initramfs. It has hence been removed entirely. The associated defconfig is deleted, as are the postbuild and postimage scripts and overlays. The patches have been relocated to the rootfs_patches directory. Also worth noting is that the machine-id setup is entirely gone from the build system.