board/MynaPlayer odyssey: fix IPv6 for the wired connection.
The default setting for generating addresses in IPv6 with NetworkManager is stable-privacy. That makes for unpredictable IPv6 addresses accross slots, and is hence entirely inacceptable. This commit changes this to be standard SLAAC without privacy instead. This also adds a device table to set the correct file permissions for the provided nmconnection file (0600 is required).