board/MynaPlayer odyssey: add pre-init script to setup overlay filesystem.
This adds a pre-init script that is ran before systemd. This scripts prepares the system so that the rootfs will be left unaltered, but combines it with a writable ext4 filesystem, where data can be stored. Both rootfs use the same overlay, which makes it possible to preserve data while overriding one of the rootfs during an upgrade.