board/MynaPlayer odyssey: remove rauc partition from genimage, and reworked the overlay filesystem creation.
The rauc partition was not needed seeing we make use of a large overlay filesystem. Remove it.

The fstab is now an empty file, so we avoid the original buildroot one.

The rauc-marking-good.service now runs mkdir -p to create the /rauc directory before it marks the slot good.

Also, the post-image script was modified to no longer create filesystems using fallocate. This was honestly dirty and added one more dependency in building MynaPlayer.

Genimage now takes care of creating the overlay filesystem instead.