board/Phoenix: added genimage configuration for PSLC flash.
The eMMC on odyssey can be configured in enhanced user area mode, which set it in PSLC (pseudo single level cell) mode. Only 1 bit per cell flash is written instead of 2 like in MLC, increasing the lifetime of the flash at the cost of capacity.

A genimage configuration is now provided to make use of this.